Use py-trello.
Use the py-trello package instead of direct HTTP.  Also set op_sys and
platform to "Unspecified" when filing the bug, in case the product is set to
try to "Detect" (which causes the API to error out if those settings are not
specified).

Includes a requirements.txt but should be packaged properly at some point.